Big refactor leveraging `ref` in actor classes and converting the interpreter and parser state to structs

+++ Source/Roton.Test/Roton/Integration/Elements/CentipedeTests.cs
@@ -118,15 +118,15 @@ public class CentipedeTests(Context context) : ElementTestFixture(context)
         var mid = Actors[seg1Index];
         var newTail = Actors[headIndex];
 
-        newHead.Follower.Should().Be(seg1Index,
+        ((int)newHead.Follower).Should().Be(seg1Index,
             "centipede head should lead its attached segment");
-        mid.Leader.Should().Be(seg2Index,
+        ((int)mid.Leader).Should().Be(seg2Index,
             "center segment should follow its head");
-        mid.Follower.Should().Be(headIndex,
+        ((int)mid.Follower).Should().Be(headIndex,
             "center segment should lead its following segment");
-        newTail.Leader.Should().Be(seg1Index,
+        ((int)newTail.Leader).Should().Be(seg1Index,
             "rear segment should follow its leader");
-        newTail.Follower.Should().Be(-1,
+        ((int)newTail.Follower).Should().Be(-1,
             "rear segment should not have a follower");
     }
