+++ Source/Roton/Emulation/Interactions/Impl/BombInteraction.cs
@@ -1,5 +1,4 @@
 ﻿using Roton.Emulation.Core;
-using Roton.Emulation.Data;
 using Roton.Emulation.Data.Impl;
 using Roton.Infrastructure.Impl;
 
@@ -16,7 +15,7 @@ public sealed class BombInteraction(IEngineAccessor engine) : IInteraction
         var actor = Engine.ActorAt(location);
         if (actor.P1 == 0)
         {
-            actor.P1 = Engine.Facts.BombCountdownStart;
+            actor.P1 = (byte)Engine.Facts.BombCountdownStart;
             Engine.UpdateBoard(location);
             Engine.SetMessage(Engine.Facts.LongMessageDuration, Engine.Alerts.BombMessage);
             Engine.PlaySound(4, Engine.Sounds.BombActivate);
